# e-commerce/internal/dto/products/productDto.go
package dto

type ProductAddRequest struct {
	Name        string  `json:"name" binding:"required"`
	Description string  `json:"description" binding:"required"`
	Price       float64 `json:"price" binding:"required" `
	Category    string  `json:"category" `
}